feat(material/legacy-progress-bar): step 1. rename directory
