+++ src/components-examples/material/progress-bar/BUILD.bazel
@@ -16,8 +16,8 @@ ng_module(
         "//src/cdk/testing",
         "//src/cdk/testing/testbed",
         "//src/material/legacy-card",
-        "//src/material/progress-bar",
-        "//src/material/progress-bar/testing",
+        "//src/material/legacy-progress-bar",
+        "//src/material/legacy-progress-bar/testing",
         "//src/material/radio",
         "//src/material/slider",
         "@npm//@angular/forms",
@@ -43,8 +43,8 @@ ng_test_library(
         ":progress-bar",
         "//src/cdk/testing",
         "//src/cdk/testing/testbed",
-        "//src/material/progress-bar",
-        "//src/material/progress-bar/testing",
+        "//src/material/legacy-progress-bar",
+        "//src/material/legacy-progress-bar/testing",
         "@npm//@angular/platform-browser-dynamic",
     ],
 )
