+++ src/components-examples/material/progress-bar/progress-bar-configurable/progress-bar-configurable-example.ts
@@ -1,6 +1,6 @@
 import {Component} from '@angular/core';
 import {ThemePalette} from '@angular/material/core';
-import {ProgressBarMode} from '@angular/material/progress-bar';
+import {ProgressBarMode} from '@angular/material/legacy-progress-bar';
 
 /**
  * @title Configurable progress-bar
