+++ src/dev-app/card/BUILD.bazel
@@ -13,7 +13,7 @@ ng_module(
         "//src/material/button",
         "//src/material/divider",
         "//src/material/legacy-card",
-        "//src/material/progress-bar",
+        "//src/material/legacy-progress-bar",
     ],
 )
