+++ src/dev-app/progress-bar/BUILD.bazel
@@ -12,7 +12,7 @@ ng_module(
     deps = [
         "//src/material/button",
         "//src/material/button-toggle",
-        "//src/material/progress-bar",
+        "//src/material/legacy-progress-bar",
         "@npm//@angular/forms",
     ],
 )
