+++ src/e2e-app/BUILD.bazel
@@ -64,9 +64,9 @@ ng_module(
         "//src/material/grid-list",
         "//src/material/icon",
         "//src/material/input",
+        "//src/material/legacy-progress-bar",
         "//src/material/list",
         "//src/material/menu",
-        "//src/material/progress-bar",
         "//src/material/progress-spinner",
         "//src/material/radio",
         "//src/material/select",
