+++ src/material-experimental/mdc-progress-bar/public-api.ts
@@ -16,4 +16,4 @@ export {
   MAT_PROGRESS_BAR_LOCATION_FACTORY,
   MAT_PROGRESS_BAR_DEFAULT_OPTIONS,
   MatProgressBarDefaultOptions,
-} from '@angular/material/progress-bar';
+} from '@angular/material/legacy-progress-bar';
