+++ src/material-experimental/mdc-progress-bar/testing/progress-bar-harness.spec.ts
@@ -1,6 +1,6 @@
 import {MatProgressBarModule} from '../index';
 import {MatProgressBarHarness} from './progress-bar-harness';
-import {runHarnessTests} from '@angular/material/progress-bar/testing/shared.spec';
+import {runHarnessTests} from '@angular/material/legacy-progress-bar/testing/shared.spec';
 
 describe('MDC-based MatProgressBarHarness', () => {
   runHarnessTests(MatProgressBarModule, MatProgressBarHarness as any);
