+++ src/material/_index.scss
@@ -92,7 +92,7 @@
 @forward './menu/menu-theme' as menu-* show menu-theme, menu-color, menu-typography;
 @forward './paginator/paginator-theme' as paginator-* show paginator-theme, paginator-color,
   paginator-typography;
-@forward './progress-bar/progress-bar-theme' as progress-bar-* show progress-bar-theme,
+@forward './legacy-progress-bar/progress-bar-theme' as progress-bar-* show progress-bar-theme,
   progress-bar-color, progress-bar-typography;
 @forward './progress-spinner/progress-spinner-theme' as progress-spinner-* show
   progress-spinner-theme, progress-spinner-color, progress-spinner-typography;
