+++ src/material/_theming.scss
@@ -26,7 +26,7 @@
 @forward './list/list-legacy-index';
 @forward './menu/menu-legacy-index';
 @forward './paginator/paginator-legacy-index';
-@forward './progress-bar/progress-bar-legacy-index';
+@forward './legacy-progress-bar/progress-bar-legacy-index';
 @forward './progress-spinner/progress-spinner-legacy-index';
 @forward './radio/radio-legacy-index';
 @forward './select/select-legacy-index';
