+++ src/material/config.bzl
@@ -40,8 +40,8 @@ entryPoints = [
     "menu/testing",
     "paginator",
     "paginator/testing",
-    "progress-bar",
-    "progress-bar/testing",
+    "legacy-progress-bar",
+    "legacy-progress-bar/testing",
     "progress-spinner",
     "progress-spinner/testing",
     "radio",
