+++ src/material/core/_core.import.scss
@@ -31,7 +31,7 @@
 @forward '../icon/icon-theme.import';
 @forward '../input/input-theme.import';
 @forward '../menu/menu-theme.import';
-@forward '../progress-bar/progress-bar-theme.import';
+@forward '../legacy-progress-bar/progress-bar-theme.import';
 @forward '../progress-spinner/progress-spinner-theme.import';
 @forward '../radio/radio-theme.import';
 @forward '../select/select-theme.import';
