+++ src/material/core/theming/_all-theme.import.scss
@@ -32,7 +32,7 @@
 @forward '../../icon/icon-theme.import';
 @forward '../../input/input-theme.import';
 @forward '../../menu/menu-theme.import';
-@forward '../../progress-bar/progress-bar-theme.import';
+@forward '../../legacy-progress-bar/progress-bar-theme.import';
 @forward '../../progress-spinner/progress-spinner-theme.import';
 @forward '../../radio/radio-theme.import';
 @forward '../../select/select-theme.import';
@@ -68,7 +68,7 @@
 @import '../../list/list-theme';
 @import '../../menu/menu-theme';
 @import '../../paginator/paginator-theme';
-@import '../../progress-bar/progress-bar-theme';
+@import '../../legacy-progress-bar/progress-bar-theme';
 @import '../../progress-spinner/progress-spinner-theme';
 @import '../../radio/radio-theme';
 @import '../../select/select-theme';
