+++ src/material/core/theming/tests/test-css-variables-theme.scss
@@ -22,7 +22,7 @@
 @use '../../../list/list-theme';
 @use '../../../menu/menu-theme';
 @use '../../../paginator/paginator-theme';
-@use '../../../progress-bar/progress-bar-theme';
+@use '../../../legacy-progress-bar/progress-bar-theme';
 @use '../../../progress-spinner/progress-spinner-theme';
 @use '../../../radio/radio-theme';
 @use '../../../select/select-theme';
