+++ src/material/core/typography/_all-typography.scss
@@ -18,7 +18,7 @@
 @use '../../list/list-theme';
 @use '../../menu/menu-theme';
 @use '../../paginator/paginator-theme';
-@use '../../progress-bar/progress-bar-theme';
+@use '../../legacy-progress-bar/progress-bar-theme';
 @use '../../progress-spinner/progress-spinner-theme';
 @use '../../radio/radio-theme';
 @use '../../select/select-theme';
