+++ src/material/legacy-progress-bar/BUILD.bazel
@@ -13,7 +13,7 @@ load(
 package(default_visibility = ["//visibility:public"])
 
 ng_module(
-    name = "progress-bar",
+    name = "legacy-progress-bar",
     srcs = glob(
         ["**/*.ts"],
         exclude = ["**/*.spec.ts"],
@@ -30,7 +30,7 @@ ng_module(
 )
 
 sass_library(
-    name = "progress_bar_scss_lib",
+    name = "legacy_progress_bar_scss_lib",
     srcs = glob(["**/_*.scss"]),
     deps = ["//src/material/core:core_scss_lib"],
 )
@@ -51,7 +51,7 @@ ng_test_library(
         exclude = ["**/*.e2e.spec.ts"],
     ),
     deps = [
-        ":progress-bar",
+        ":legacy-progress-bar",
         "//src/cdk/testing/private",
         "@npm//@angular/platform-browser",
     ],
