+++ src/material/legacy-progress-bar/_progress-bar-legacy-index.scss
