+++ src/material/legacy-progress-bar/_progress-bar-theme.import.scss
