+++ src/material/legacy-progress-bar/_progress-bar-theme.scss
@@ -68,7 +68,7 @@
 
 @mixin theme($theme-or-color-config) {
   $theme: theming.private-legacy-get-theme($theme-or-color-config);
-  @include theming.private-check-duplicate-theme-styles($theme, 'mat-progress-bar') {
+  @include theming.private-check-duplicate-theme-styles($theme, 'mat-legacy-progress-bar') {
     $color: theming.get-color-config($theme);
     $density: theming.get-density-config($theme);
     $typography: theming.get-typography-config($theme);
