+++ src/material/legacy-progress-bar/progress-bar-module.ts
