+++ src/material/legacy-progress-bar/progress-bar.e2e.spec.ts
