+++ src/material/legacy-progress-bar/progress-bar.html
