+++ src/material/legacy-progress-bar/progress-bar.scss
