+++ src/material/legacy-progress-bar/progress-bar.spec.ts
