+++ src/material/legacy-progress-bar/public-api.ts
