+++ src/material/legacy-progress-bar/testing/BUILD.bazel
@@ -26,7 +26,7 @@ ng_test_library(
         ":testing",
         "//src/cdk/testing",
         "//src/cdk/testing/testbed",
-        "//src/material/progress-bar",
+        "//src/material/legacy-progress-bar",
         "@npm//@angular/platform-browser",
     ],
 )
@@ -40,7 +40,7 @@ ng_test_library(
     deps = [
         ":harness_tests_lib",
         ":testing",
-        "//src/material/progress-bar",
+        "//src/material/legacy-progress-bar",
     ],
 )
