+++ src/material/legacy-progress-bar/testing/progress-bar-harness.spec.ts
@@ -1,5 +1,5 @@
-import {MatProgressBarModule} from '@angular/material/progress-bar';
-import {runHarnessTests} from '@angular/material/progress-bar/testing/shared.spec';
+import {MatProgressBarModule} from '@angular/material/legacy-progress-bar';
+import {runHarnessTests} from '@angular/material/legacy-progress-bar/testing/shared.spec';
 import {MatProgressBarHarness} from './progress-bar-harness';
 
 describe('MatProgressBarHarness', () => {
