+++ src/material/legacy-progress-bar/testing/progress-bar-harness.ts
