+++ src/material/legacy-progress-bar/testing/public-api.ts
